pyproject.toml
[project]
name = "attobolt"
version = "0.1.0"
description = "Slack bot that gives you a Claude Code session from anywhere"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "slack-bolt>=1.21.0",
    "python-dotenv>=1.0.0",
    "watchfiles>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["attobolt"]
